redesign of UpcomingEventsFragment

+++ app/src/main/java/ru/myitschool/nasa_bootcamp/data/dto/spaceX/cores/Core.kt
@@ -9,7 +9,7 @@ class Core(
     @field:SerializedName("block") val block: Int?,
     @field:SerializedName("status") val status: String?,
     @field:SerializedName("original_launch") val original_launch: String,
-    @field:SerializedName("original_launch_unix") val original_launch_unix: Int,
+    @field:SerializedName("original_launch_unix") val original_launch_unix: Long,
     @field:SerializedName("missions") val missions: List<Mission>,
     @field:SerializedName("reuse_count") val reuse_count: Int,
     @field:SerializedName("water_landing") val water_landing: Boolean,
