+++ app/src/main/java/ru/myitschool/nasa_bootcamp/data/model/HistoryModel.kt
@@ -3,6 +3,6 @@ package ru.myitschool.nasa_bootcamp.data.model
 data class HistoryModel(
     val id: Int,
     val title: String,
-    val event_date_utc: Int,
+    val event_date_utc: Long,
     val details: String
 )
